agent/requirements.txt
livekit-agents[mcp]>=0.8.0
livekit-plugins-openai>=0.8.0
livekit-plugins-azure>=0.3.0
livekit-plugins-silero>=0.6.0
livekit-plugins-assemblyai>=0.1.0
livekit-plugins-cartesia>=1.1.0
python-dotenv>=1.0.0
supabase>=2.0.0
pytz>=2024.1
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0